organizing and add hook for gallery json
