+++ src/components/Gallery/PieceCard/index.css
@@ -1,9 +1,3 @@
-/* Gallery.css */
-.card-container {
-    display: flex;
-    flex-wrap: wrap;
-}
-
 .card {
     border: 1px solid #ccc;
     border-radius: 8px;
