+++ src/pages/gallery/index.css
@@ -1,0 +1,4 @@
+.card-container {
+    display: flex;
+    flex-wrap: wrap;
+}
