+++ src/utils/galleryPiecesMock.ts
